% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_toml.R
\name{read_toml}
\alias{read_toml}
\alias{parse_toml}
\title{Convert a TOML string to an R list}
\usage{
parse_toml(x, flatten = TRUE)

read_toml(x, flatten = TRUE)
}
\arguments{
\item{x}{A length-one character vector, either a path to a file for
\code{read_toml()} or a TOML document for \code{parse_toml()}}

\item{flatten}{Convert single-type arrays to a vector, default \code{TRUE}}
}
\value{
A list
}
\description{
Convert a TOML string to an R list
}
