% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoml-package.R
\docType{package}
\name{rtoml-package}
\alias{rtoml}
\alias{rtoml-package}
\title{rtoml: Parse TOML files}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Paulin \email{paulin.mark@googlemail.com}

}
\keyword{internal}
