% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_toml.R
\name{read_toml}
\alias{read_toml}
\alias{read_toml_str}
\title{Convert a TOML string to an R list}
\usage{
read_toml_str(x)

read_toml(x)
}
\arguments{
\item{x}{A length-one character vector}
}
\value{
A list
}
\description{
Convert a TOML string to an R list
}
