% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoml-package.R
\docType{package}
\name{rtoml-package}
\alias{rtoml}
\alias{rtoml-package}
\title{rtoml: Parse TOML files}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarkPaulin/rtoml}
  \item Report bugs at \url{https://github.com/MarkPaulin/rtoml/issues}
}

}
\author{
\strong{Maintainer}: Mark Paulin \email{paulin.mark@googlemail.com}

}
\keyword{internal}
